Several changes - Deleted ErrorView which was completely unused - Renamed Views -> UserInterface, grouped some reusable code under "UserInterface/Components" - Renamed DataModels -> Model, moved everything from Components&Extensions/Components to this location - Moved Extensions folder out to the main tree - Removed a comment from AppModel, line 208
