+++ AnAppleADay/Extensions/Entity.swift
