+++ AnAppleADay/Model/OnboardingParameters.swift
