+++ AnAppleADay/UserInterface/GenerateModelView.swift
